Aggiunto/sistemato errore sessione scadute

+++ php/AdminModAnnunci.php
@@ -21,53 +21,59 @@ menu($page);
 
 # ------------------------------------------------------
 
-echo"<div id='contenuto'>";
+if(isset($_SESSION['login'])){
+	echo"<div id='contenuto'>";
 
-//delete action
-if(isset($_POST['delete'])) {
-	$ad = $_POST['delete'];
+	//delete action
+	if(isset($_POST['delete'])) {
+		$ad = $_POST['delete'];
 
-	$delete = "DELETE FROM Annunci WHERE Codice='".$ad."'";
+		$delete = "DELETE FROM Annunci WHERE Codice='".$ad."'";
 
-	if(mysqli_query(openDB(), $delete)) {
-		echo "<div class='successMsg'>Rimozione dell'annuncio avvenuto con successo!</div>";
+		if(mysqli_query(openDB(), $delete)) {
+			echo "<div class='successMsg'>Rimozione dell'annuncio avvenuto con successo!</div>";
+		}
+		else {
+			echo "<div class='errorMsg'>Problemi con l'eliminizione dell'annuncio</div>";
+		}
 	}
+
+	$result = mysqli_query(openDB(), "SELECT * FROM Annunci ORDER BY Titolo");
+
+	if(mysqli_num_rows($result) == 0) {
+		echo "<div id='NoData'> Nessun annuncio ancora inserito. </div>";
+	}
+
 	else {
-		echo "<div class='errorMsg'>Problemi con l'eliminizione dell'annuncio</div>";
+		echo "<h3>Ecco gli annunci che sono stati pubblicati:</h3>
+			<ul>";
+		
+		while($row = $result->fetch_array(MYSQLI_ASSOC)) {
+			echo "<li class='ResAnn'>
+					<ul>
+						<li id='title'> ".$row['Titolo'].";</li>
+						<li><strong>Tipologia</strong>: ".$row['Tipologia'].";</li>
+						<li><strong>Con contratto</strong>: ".$row['Contratto'].";</li>
+						<li><strong>Con orario</strong>: ".$row['Orario'].";</li>
+						<li><strong>Dall'azienda</strong>: ".$row['Azienda'].";</li>
+						<li><strong>Il giorno</strong>: ".$row['Data'].";</li>
+						<li><strong>Descrizione</strong>: ".$row['Descrizione'].";</li>
+						
+							<div id='options'>
+								<form method='post' action='AdminModAnnunci.php'>
+					            	<button value=".$row['Codice']." name='delete'>Rimuovi</button>
+				            	</form>
+			            	</div>
+					</ul>
+				</li>";
+		}
+		echo"</ul>";
 	}
 }
-
-$result = mysqli_query(openDB(), "SELECT * FROM Annunci ORDER BY Titolo");
-
-if(mysqli_num_rows($result) == 0) {
-	echo "<div id='NoData'> Nessun annuncio ancora inserito. </div>";
-}
-
-else {
-	echo "<h3>Ecco gli annunci che sono stati pubblicati:</h3>
-		<ul>";
-	
-	while($row = $result->fetch_array(MYSQLI_ASSOC)) {
-		echo "<li class='ResAnn'>
-				<ul>
-					<li id='title'> ".$row['Titolo'].";</li>
-					<li><strong>Tipologia</strong>: ".$row['Tipologia'].";</li>
-					<li><strong>Con contratto</strong>: ".$row['Contratto'].";</li>
-					<li><strong>Con orario</strong>: ".$row['Orario'].";</li>
-					<li><strong>Dall'azienda</strong>: ".$row['Azienda'].";</li>
-					<li><strong>Il giorno</strong>: ".$row['Data'].";</li>
-					<li><strong>Descrizione</strong>: ".$row['Descrizione'].";</li>
-					
-						<div id='options'>
-							<form method='post' action='AdminModAnnunci.php'>
-				            	<button value=".$row['Codice']." name='delete'>Rimuovi</button>
-			            	</form>
-		            	</div>
-				</ul>
-			</li>";
-	}
-	echo"</ul>";
-
+else{
+    echo " <div id='contenuto'>
+	           <div class='errorMsg'>Sessione scaduta, procedere con la riutenticazione.</div>
+	       </div>";
 }
 
 # ------------------------------------------------------
